[javalin] Move compression config to http config (#2002)
