+++ javalin/src/main/java/io/javalin/config/HttpConfig.kt
@@ -1,8 +1,11 @@
 package io.javalin.config
 
+import io.javalin.compression.Brotli
+import io.javalin.compression.CompressionStrategy
+import io.javalin.compression.Gzip
 import io.javalin.http.ContentType
 
-class HttpConfig {
+class HttpConfig(private val cfg: JavalinConfig) {
     //@formatter:off
     @JvmField var generateEtags = false
     @JvmField var prefer405over404 = false
@@ -10,4 +13,27 @@ class HttpConfig {
     @JvmField var defaultContentType = ContentType.PLAIN
     @JvmField var asyncTimeout = 0L
     //@formatter:on
+
+    fun custom(compressionStrategy: CompressionStrategy) {
+        cfg.pvt.compressionStrategy = compressionStrategy
+    }
+
+    @JvmOverloads
+    fun brotliAndGzip(gzipLevel: Int = 6, brotliLevel: Int = 4) {
+        cfg.pvt.compressionStrategy = CompressionStrategy(Brotli(brotliLevel), Gzip(gzipLevel))
+    }
+
+    @JvmOverloads
+    fun gzipOnly(level: Int = 6) {
+        cfg.pvt.compressionStrategy = CompressionStrategy(null, Gzip(level))
+    }
+
+    @JvmOverloads
+    fun brotliOnly(level: Int = 4) {
+        cfg.pvt.compressionStrategy = CompressionStrategy(Brotli(level), null)
+    }
+
+    fun none() {
+        cfg.pvt.compressionStrategy = CompressionStrategy(null, null)
+    }
 }
